Bug Fixes
In the game.js file, "Discord.MessageEmbed ()" is now a comment because the method was returning errors, so the messages that came out when winning or losing are now gone, it returns another error that does not affect at all, so you can keep playing.
Some modules that were in bot.js were also moved into the files that pressed them.
(now the music does not play in a loop)